Add files via upload
The file YTharvest.py consists of python code, has many functions capable of extracting the data from YouTube, like Channel details, Video details and Comment details with the help of Channel ID taken directly from YouTube. This file has a MYSQL connection to store the extracted data in local database. And also this file contains the Streamlit connection to display the data according to the SQL queries included in the code.